fix: Do not enable nvim complete
It was overlapping with nvim-cmp, causing different menus sometimes and actually selecting an entry to not work (in the nvim menu).

The built-in menu is also vastly inferior: No colors → less overview, less/worse additional-info popups (method signatures, class headers and so on). It also doesn't support (or at least I didn't integrate it with) luasnip.

So glad I finally found out what this issue was so I can just nuke it. We learn: Don't do recommended updating configuration unless you understand what it's doing!