fix: Increase sign column for correct window on LspAttach
Previously, I increased (=set to "yes:3") the signcolumn for "whatever was active"; now, I do it to whatever the LSP attached to; since signcolumn is a window option, but LSP clients attach to a buffer, we lookup the window(s) the buffer is currently in.

This is still an issue, because the buffers in a window can change, and that isn't accounted for at all; also adding a buffer that already has an LSP client attached won't change the window anymore, since LspAttach won't fire again.